[chores] Moved shared VPN backend logic to BaseVpnBackend
